requirements.txt
python-telegram-bot==21.6
supabase==2.10.0
google-generativeai==0.8.3
python-dotenv==1.0.1
Pillow==10.4.0
